+++ pulp/0.2/label_balance_verts.cpp
@@ -54,34 +54,34 @@ using namespace std;
  ########:: ##:::: ##: ########::::::. ###:::: ########: ##:::. ##:::: ##::::
 ........:::..:::::..::........::::::::...:::::........::..:::::..:::::..:::::
 */
-void label_balance_verts(pulp_graph_t& g, int num_parts, int* parts,
-  int vert_outer_iter, int vert_balance_iter, int vert_refine_iter,
+void label_balance_verts(pulp_graph_t& g, int64_t num_parts, int64_t* parts,
+  int64_t vert_outer_iter, int64_t vert_balance_iter, int64_t vert_refine_iter,
   double vert_balance)
 {
-  int num_verts = g.n;
-  int* part_sizes = new int[num_parts];
+  int64_t num_verts = g.n;
+  int64_t* part_sizes = new int64_t[num_parts];
  
   for (int i = 0; i < num_parts; ++i)
     part_sizes[i] = 0;
 
   double avg_size = num_verts / num_parts;
-  int num_swapped_1 = 0;
-  int num_swapped_2 = 0;
+  int64_t num_swapped_1 = 0;
+  int64_t num_swapped_2 = 0;
   double max_v;
   double running_max_v = (double)num_verts;
 
-  int* queue = new int[num_verts*QUEUE_MULTIPLIER];
-  int* queue_next = new int[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue_next = new int64_t[num_verts*QUEUE_MULTIPLIER];
   bool* in_queue = new bool[num_verts];
   bool* in_queue_next = new bool[num_verts];
-  int queue_size = num_verts;
-  int next_size = 0;
-  int t = 0;
-  int num_tries = 0;
+  int64_t queue_size = num_verts;
+  int64_t next_size = 0;
+  int64_t t = 0;
+  int64_t num_tries = 0;
 
 #pragma omp parallel
 {
-  int* part_sizes_thread = new int[num_parts];
+  int64_t* part_sizes_thread = new int64_t[num_parts];
   for (int i = 0; i < num_parts; ++i) 
     part_sizes_thread[i] = 0;
 
@@ -99,9 +99,9 @@ void label_balance_verts(pulp_graph_t& g, int num_parts, int* parts,
   double* part_counts = new double[num_parts];
   double* part_weights = new double[num_parts];
 
-  int thread_queue[ THREAD_QUEUE_SIZE ];
-  int thread_queue_size = 0;
-  int thread_start;
+  int64_t thread_queue[ THREAD_QUEUE_SIZE ];
+  int64_t thread_queue_size = 0;
+  int64_t thread_start;
 
   for (int p = 0; p < num_parts; ++p)
   {        
@@ -127,29 +127,29 @@ while(t < vert_outer_iter)
   next_size = 0;
 }
 
-  int num_iter = 0;
+  int64_t num_iter = 0;
   while (/*swapped &&*/ num_iter < vert_balance_iter)
   {
 #pragma omp for schedule(guided) reduction(+:num_swapped_1) nowait
     for (int i = 0; i < queue_size; ++i)
     {
-      int v = queue[i];
+      int64_t v = queue[i];
       in_queue[v] = false;
-      int part = parts[v];
+      int64_t part = parts[v];
       for (int p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
       unsigned out_degree = out_degree(g, v);
-      int* outs = out_vertices(g, v);
+      int64_t* outs = out_vertices(g, v);
       for (unsigned j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
-        int part_out = parts[out];
+        int64_t out = outs[j];
+        int64_t part_out = parts[out];
         part_counts[part_out] += out_degree(g, out);
         //part_counts[part_out] += 1.0;//out_degree(g, out);
       }
       
-      int max_part = part;
+      int64_t max_part = part;
       double max_val = 0.0;
       for (int p = 0; p < num_parts; ++p)
       {
@@ -233,7 +233,7 @@ while(t < vert_outer_iter)
 #if VERBOSE
     printf("%d\n", num_swapped_1);
 #endif
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
     bool* temp_b = in_queue;
@@ -267,23 +267,23 @@ while(t < vert_outer_iter)
 #pragma omp for schedule(guided) reduction(+:num_swapped_2) nowait  
     for (int i = 0; i < queue_size; ++i)
     {
-      int v = queue[i];
+      int64_t v = queue[i];
       in_queue[v] = false;
       for (int p = 0; p < num_parts; ++p)
         part_counts[p] = 0;
 
-      int part = parts[v];
+      int64_t part = parts[v];
       unsigned out_degree = out_degree(g, v);
-      int* outs = out_vertices(g, v);
+      int64_t* outs = out_vertices(g, v);
       for (unsigned j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
-        int part_out = parts[out];
+        int64_t out = outs[j];
+        int64_t part_out = parts[out];
         part_counts[part_out]++;
       }
 
-      int max_part = -1;
-      int max_count = -1;
+      int64_t max_part = -1;
+      int64_t max_count = -1;
       for (int p = 0; p < num_parts; ++p)
         if (part_counts[p] > max_count)
         {
@@ -358,7 +358,7 @@ while(t < vert_outer_iter)
 #if VERBOSE
     printf("%d\n", num_swapped_2);
 #endif
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
     bool* temp_b = in_queue;
@@ -427,11 +427,11 @@ while(t < vert_outer_iter)
 ........:::..:::::..::........::::::::...:::::........::..:::::..:::::..:::::
 */
 void label_balance_verts_weighted(
-  pulp_graph_t& g, int num_parts, int* parts,
-  int vert_outer_iter, int vert_balance_iter, int vert_refine_iter,
+  pulp_graph_t& g, int64_t num_parts, int64_t* parts,
+  int64_t vert_outer_iter, int64_t vert_balance_iter, int64_t vert_refine_iter,
   double vert_balance)
 {
-  int num_verts = g.n;
+  int64_t num_verts = g.n;
   long* part_sizes = new long[num_parts];
 
   bool has_vwgts = (g.vertex_weights != NULL);
@@ -442,19 +442,19 @@ void label_balance_verts_weighted(
     part_sizes[i] = 0;
 
   double avg_size = (double)g.vertex_weights_sum / (double)num_parts;
-  int num_swapped_1 = 0;
-  int num_swapped_2 = 0;
+  int64_t num_swapped_1 = 0;
+  int64_t num_swapped_2 = 0;
   double max_v;
   double running_max_v = (double)num_verts;
 
-  int* queue = new int[num_verts*QUEUE_MULTIPLIER];
-  int* queue_next = new int[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue_next = new int64_t[num_verts*QUEUE_MULTIPLIER];
   bool* in_queue = new bool[num_verts];
   bool* in_queue_next = new bool[num_verts];
-  int queue_size = num_verts;
-  int next_size = 0;
-  int t = 0;
-  int num_tries = 0;
+  int64_t queue_size = num_verts;
+  int64_t next_size = 0;
+  int64_t t = 0;
+  int64_t num_tries = 0;
 
 #pragma omp parallel
 {
@@ -479,9 +479,9 @@ void label_balance_verts_weighted(
   double* part_counts = new double[num_parts];
   double* part_weights = new double[num_parts];
 
-  int thread_queue[ THREAD_QUEUE_SIZE ];
-  int thread_queue_size = 0;
-  int thread_start;
+  int64_t thread_queue[ THREAD_QUEUE_SIZE ];
+  int64_t thread_queue_size = 0;
+  int64_t thread_start;
 
   for (int p = 0; p < num_parts; ++p)
   {        
@@ -507,34 +507,34 @@ while(t < vert_outer_iter)
   next_size = 0;
 }  
 
-  int num_iter = 0;
+  int64_t num_iter = 0;
   while (/*swapped &&*/ num_iter < vert_balance_iter)
   {
 #pragma omp for schedule(guided) reduction(+:num_swapped_1) nowait
     for (int i = 0; i < queue_size; ++i)
     {
-      int v = queue[i];
+      int64_t v = queue[i];
       in_queue[v] = false;
-      int part = parts[v];
-      int v_weight = 1;
+      int64_t part = parts[v];
+      int64_t v_weight = 1;
       if (has_vwgts) v_weight = g.vertex_weights[v];
 
       for (int p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
       unsigned out_degree = out_degree(g, v);
-      int* outs = out_vertices(g, v);
-      int* weights = out_weights(g, v);
+      int64_t* outs = out_vertices(g, v);
+      int64_t* weights = out_weights(g, v);
       for (unsigned j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
-        int part_out = parts[out];
+        int64_t out = outs[j];
+        int64_t part_out = parts[out];
         double weight_out = 1.0;
         if (has_ewgts) weight_out = (double)weights[j];
         part_counts[part_out] += (double)out_degree(g, out)*weight_out;
       }
       
-      int max_part = part;
+      int64_t max_part = part;
       double max_val = 0.0;
       for (int p = 0; p < num_parts; ++p)
       {
@@ -618,7 +618,7 @@ while(t < vert_outer_iter)
 #if VERBOSE
     printf("%d\n", num_swapped_1);
 #endif
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
     bool* temp_b = in_queue;
@@ -652,29 +652,29 @@ while(t < vert_outer_iter)
 #pragma omp for schedule(guided) reduction(+:num_swapped_2) nowait  
     for (int i = 0; i < queue_size; ++i)
     {
-      int v = queue[i];
+      int64_t v = queue[i];
       in_queue[v] = false;      
-      int part = parts[v];
-      int v_weight = 1;
+      int64_t part = parts[v];
+      int64_t v_weight = 1;
       if (has_vwgts) v_weight = g.vertex_weights[v];
 
       for (int p = 0; p < num_parts; ++p)
         part_counts[p] = 0;
 
       unsigned out_degree = out_degree(g, v);
-      int* outs = out_vertices(g, v);
-      int* weights = out_weights(g, v);
+      int64_t* outs = out_vertices(g, v);
+      int64_t* weights = out_weights(g, v);
       for (unsigned j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
-        int part_out = parts[out];
-        int out_weight = 1;
+        int64_t out = outs[j];
+        int64_t part_out = parts[out];
+        int64_t out_weight = 1;
         if (has_ewgts) out_weight = weights[j];
         part_counts[part_out] += out_weight;
       }
 
-      int max_part = -1;
-      int max_count = -1;
+      int64_t max_part = -1;
+      int64_t max_count = -1;
       for (int p = 0; p < num_parts; ++p)
         if (part_counts[p] > max_count)
         {
@@ -749,7 +749,7 @@ while(t < vert_outer_iter)
 #if VERBOSE
     printf("%d\n", num_swapped_2);
 #endif
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
     bool* temp_b = in_queue;
